Spec Query#group_names_for_entity_and_acting_as
There is a naive hash that translates a role name to a group name.
I want to expose a query method to retrieve that information; This
information will most certainly make it into a database for greater
configuration options.

Also, consider the fact that the Sipity::Commands::PermissionCommands
module will need to be changed and the underlying spec would benefit
from a tweak as well, saying that the Sipity::Queries::GroupQueries
is allowed to receive #group_names_for_entity_and_roles and return a
specified value.

See Sipity::Commands::PermissionCommands#grant_groups_permission_to_entity_for_role

[skip ci]

Refines to #52