Adding mutant
I want to begin exploring mutation coverage. There are a few critical
paths that I want to make sure are playing well, and mutation testing
is the path to verify.

See the following blog post for reasoning:

* http://blog.arkency.com/2015/04/why-i-want-to-introduce-mutation-testing-to-the-rails-event-store-gem/

To use:

```console
$ RAILS_ENV=test bundle exec mutant -r ./config/environment \
  --use rspec Sipity::Forms::ProcessingForm
```

WARNING: The above test takes a long time (8 minutes or so)