Reducing default pagination from 15 to 8
There are underlying slowness on the index page due to the flexible data
structure of Sipity and not a direct consideration for the data-
structure necessary for rendering an index view instead of a show page.

This is a quick fix meant to address DLTP-967. If the reduced pagination
does not improve the apparent responsiveness, we'll need to look into
additional measures (e.g. define the data-structure and populate that
with query results)