Updating Rubocop gem
To address a security vulnerability, I bumped the gem to the nearest
version that resolved the vulnerability.

Addressing some of the historical oddities of the `.hound.yml` file,
I leveraged Rubocop's inherit_from declarations. In essence deprecating
the `.hound.yml` file in favor of the conventional `.rubocop.yml` file.

Then, as there is a version update, I opted to resolve a few low-hanging
fruit but for expendiency added the `.rubocop_todo.yml` to say these are
things to work through.