Updating fixtures to reference records
As previously, the fixtures and specs assumed that before each spec
ran, the database table's auto-incrementing columns each reset to start
from 1. That assumption holds true with sqlite3. However that does not
hold true for mysql or postgresql. I aspire to switch the test
environment from `sqlite` to `mysql` and to align the test database
with the production and development databases. As I used AREL and
ActiveRecord, the upstream abstractions helped ensure "common enough"
behavior. However as Rails continues to introduce more specific
migration idioms, that "common enough" no longer remains adequate.

This change fixes both fixtures and specs that relied on set record
identifiers (a short-cut that I took during the heart of development
of Sipity but one that impedes longer term maintenance and runs against
what I consider to be best practices).