Updating schema to address Rails 5.2 realities
Rails does not adhere to Semantic versioning. Instead they have a
policy in which the minor version can introduce breaking changes. This
usually means prior patch versions will include deprecation notices.

However, in the case of the db/schema.rb file this is not something
that is necessarily vetted. By default the db/schema.rb is only updated
after a migration that has not yet been run is run. When I drop the
local database and run the rake tasks, many things fall apart. This
update address at least the test environment.