Removing the "Read More" abstract link
Prior to this commit, and since the v2019.6 release, the "Read More"
link broke; The JS function rendered the "Read More" link however the
text which should have been hidden still displayed on the page,
overlaying the content below the abstract.

I did some debugging and was unable to find a reasonable solution. In
consultation with Shari, she agreed that we could remove the "Read More"
link and instead display the entire abstract.