Restoring "Add another Person" button
In my haste to push out a fix to the broken "Read More" / "Read Less"
behavior for abstracts, I removed code that fired the "Add another
Person" button.

This change fixes that.

In addition, in running the `rake bootstrap` command to have a fresh
environment, the `db/schema.rb` update came along. It is a rather
innocuous add.