Adding tooling for testing the WEBHOOK Callback
One "hole" in the system is that the WEBHOOK Callback is not fully
tested. This is somewhat difficult in the feature specs; I need a server
running to receive the request; that server may a background server for
Capybara (and not have the same port information).

I have added a few tasks that that can verify the WEBHOOK works within
a running development environment.

I also updated the [documentation][1] to include this task

[1]: https://docs.google.com/document/d/1u97Ncwkg4KjKdaa53vTXaZVpTRZO2KCPYmjNVEwa7_8/edit#heading=h.lqfwb5y2t3zz