Adjusting requests outside pagination range
Prior to this commit, when requesting a paginated list of works, the
requester would always get a 200 HTTP response when the requested page
was within the total pages for the pagination OR when the requested
page was outside the pagination range. The response document would
have an empty array of works and pagination links (self, first, last,
prev, next) as per the JSON+API specification.

In the case of one API integration, the requester appeared to follow
the following logic:

```
let page = 1
request page
while response status is 200
  for each element in work array
    run some process
  end for
  increment page by 1
  request page
end while
```

The process did not appear to check that the response had any elements
in the array and stop once it encountered an empty array.

With this commit, the pagination logic has changed a bit:

* When you request page 1, you will get a 200 response.
* When you request a page within the pagination window (e.g. between 1
  and total number of pages) you will get a 200 response.
* When you request a page outside the pagination window, you will get
  a 404 response