Removing ETD_INTEGRATORS from ingested state
Prior to this commit, the OIT's integration would see ETD works in the
Ingesting and Ingested stated. This meant that all objects that had
been approved by the Grad School were processed each time. This
resulted in at least 170 pages of data to process, and was growing each
year.

Per Grad School instructions, I'm removing access to the
ETD_INTEGRATORS for ingesting and ingested state.

From Shari:

> I don't recall the specifics. I know that it's helpful for us to have
> updates pushed to OnBase if we correct the title or advisor on the
> Sipity record before we approve it (for example). If we could set
> OnBase to stop updating a given record once the Sipity State hits
> "ingesting" that would be the best of all worlds for us.

Upon further discussion, the OnBase process has detected things stuck
in the "ingesting" state. In conversation with Shari, we agreed that
we would preserve permission to OnBase for "ingesting."

In addition, adding service class to assist in revoking permissions for
a given role.

This relates and builds on https://github.com/ndlib/sipity/pull/1227