Enabling Grad School editing of submission date
Prior to this commit, Sipity stamped the etd_submission_date at the
point when the student submitted their work. Sipity did not then
provide a means of updating etd_submission_date (aside from using
SQL or the Rails console).

With this commit, I added a form in which the grad school can edit the
etd_submission_date.

This feature was prompted in an email thread with Shari.