Adding full NOIDs integration
Prior to this commit, Sipity did not have a means of testing against a
running [our NOIDs server](https://github.com/ndlib/noids). This
commit refactors the specs and minter implementation to make it
possible to run an integration test against a "live" NOIDs server.

The refactor preserves the lambda method signature in the [application
config][1] for the configured default_pid_minter.

Dependent on ndlib/noids_client#5

[1]:./config/application.rb