Adding ability to swap method used for NetID read
We encountered the following error:

```shell
OpenSSL::SSL::SSLError
SSL_connect returned=1 errno=0 state=error: certificate verify failed
```

Triaging the problem, we discovered an intermittent issue that we are
working to resolve. In the mean time, I worked through this commit.

First, I adjusted the system to provide two different reader methods. I
did not want to lose sight of the "when things are working" state. I
also wanted to draw attention to the "when things aren't working" state
and how we solved that.

This commit makes the upcoming change easier. By design, I'm exposing
the workaround as an option, but with this commit continuing to use the
desired method.