Adding NightlyJob exception handling
Sipity is already raising an individiual exception when it fails to
"ingest" a single work.  What is missing, however, is a greater
context (and sharing of what might be happening).

In adding this new bit of logic, my hope is to include a more verbose
exception that explains: "Yup, there's an error in the bulk job, and as
a result the work's status likely didn't change.  So go look at the
details of that exception."

In other words, the presence of this newly added error says: "Hey, Batch
Ingest had a problem."