Remove Dashboard
Remove dashboard controller and spec.

Purpose of controller is for a user to see all of their possible actions
in one view. However dashboard is basically another version
of the ETD area's home page when there is only one work area. The effort
to keep the functionality in sync between the views is unjustified at
this point.

If additional work areas are needed in the future, this functionality
can be restored.