Interstial change to add two attributes to dashboard
I'm not certain that this will work, but I want to capture the logic
that I've done to generate the joins.

The reason that this is not quite right is that I need to add the change
into given `criteria` AND `scope_proxied_objects_for_the_user_and_proxy_for_type` method.

The `criteria` object defines what are the sort columns and query
strings; so I need to expose the field sort for that.  The
`scope_proxied_objects_for_the_user_and_proxy_for_type` creates the
initial AREL scope from which we then build out additional queries.

So, while this "kind of works" and "kind of" addresses the
situation (e.g., the product owner wants to see the author and the
submission date in the dashboard and the dashboard is too slow), it's
not yet done.