Moving methods onto search criteria
This is a bit of an odd duck, the criteria parameters are there to help
consolidate the logic necessary to run a search/filter/sort query for
the dashboard.  So there are two inter-related things:

1) The Active Record scope object
2) The criteria parameters

And one updates the other; regardless of how I do this.  I could extract
the logic to happen within the criteria (and am tempted given that I'm
now blending SQL logic into that criteria).

Ultimately, the goal is to allow for the querying of the dashboard to:

1) include a sort on an additional attribute (e.g. submission_date)
2) reduce the number of ActiveRecord queries necessary for building out the
   dashboard

This change iterates in that direction.

What remains?

- [ ] adding the sort criteri
- [ ] updating the templates to render those fields
- [ ] ripping out some of the Curly view object’s antics
- [ ] perhaps removing an `includes` directive on the dashboard's query