Improving performance of dashboard response
Prior to this commit, for each work on the dashboard, we'd query the
additional parameters for the program names.  With this change, we're
defaulting the dashboard query to add the program_name(s) to the initial
query result set.  This removes one of the conditions of N+1 queries.

One small concession: the SQL concatentation is with a ", " instead of
using Rails's `to_sentence`.  The result is that the array `["A", "B",
"C"]` goes from "A, B, and C" to "A, B, C".