Having ShowPresenter specify additional_attributes
Prior to this commit, I made the assumption that the ShowPresenter would
defer to the `SearchCriteriaForWorksParameter` for the additional
attributes.

However, as part of the PR, LaRita recommended that each context should
specify it's additional attributes.  Which is spot on.

This change now, in essence, makes the prior changes an additive change
that by default doesn't impact other implementations.  And instead
updates the dashboard rendering behavior to be the only place where the
additional attributes of "author_name", "etd_submission_date", and
"program_name" might be applied.