Parameterizing list for states to allow non-terminal
This commit is a refactoring of a query, to allow for more immediate and
simplified AREL query chaining.

The goal of this change is to introduce a means for generating a list of
non-terminal states for the given work area.  Non-terminal states are
those states that have actions that can transition them out of the given
state.  A terminal state is an a state in which there are no actions
that will transition the object out of it's current state.