Adding default omit terminal actions (#1292)
* Interstitial commit to query multiple states

With this commit, you can now select and filter on multiple states on
the dashboard.

* Switching work area interface

This aligns with the emerging changes for the
`SearchCriteriaForWorksParameter` interface (e.g., moving from single
value to multiple value).

* Updating work_area to be an "first set" attribute

Prior to this commit, the order of assignment was somewhat arbitrary.
And as I want to amend the default_processing_states to start with
non-terminal states, I need to ensure that work_area is set before the
others (so that I can query the work area's non-terminal states).

* Renaming search/filter parameter to plural

Prior to this change, the parameter name presumed a singular value.
With this change, the parameter name implies plurality.

As part of the refactor, I've removed the singular name (which could
remain if we alias plural to singular, but instead I chose to remove it
for clarity).

* Parameterizing list for states to allow non-terminal

This commit is a refactoring of a query, to allow for more immediate and
simplified AREL query chaining.

The goal of this change is to introduce a means for generating a list of
non-terminal states for the given work area.  Non-terminal states are
those states that have actions that can transition them out of the given
state.  A terminal state is an a state in which there are no actions
that will transition the object out of it's current state.

* Removing "all states" from select list

In local testing, the Rails logic around the input prompt and
multi-select can create some inconsistent behavior.

For example, if I selected only "All States", then ran the search, the
returned page would not indicate that "All States" was selected.  In
other words, the select box would have nothing selected.  I found that
this created an odd visual state.

Thinking about this it makes sense; what does it mean to have a
multi-select and then someone selects "All States" and "New"?  Which do
we honor in the filter?

My response is to remove the "All States" prompt/placeholder thus
showing specifically which states you've selected.  I suppose if you
select NONE of them, the inner logic will render all of the states.

Addresses - https://jira.library.nd.edu/browse/CURATE-289