Adding service to change creating user (#1297)
Based on conversations with LaRita, we've now had two cases where
someone started a submission on behalf of another person.  This service
attempts to codify how to change the user.

Note: as an administrative function for which there's only been 2 uses,
I have written "happy path" specs.  This means, I'm demonstrating that I
can successfully transfer the "creating_user" role from one user to
another.

There are bits of the audit that I'm not testing, as that would require
a few more expensive tests to run.