Export attachments as json (#1294)
* Export attachments as json

Rather than sending full attachment files via HTTP request during
export, create a json file of the attachment location information, and
add that file to the request.

* Remove put_file expectation

Remove expectation in specs due to removal of method `put_file`.`

* Rename search title

Prior title was misleading.

* Rename attachments file

Since the attachments.json file is created in a work-specific directory,
a work ID is an unnecessary part of the file name. Removing it
simplifies the ingest logic.